BI-14526: Tidy up tests and update makefile/pom
* Makefile and Pom updated to have most recent plugin config and build functionality, similar to acsps and primary-search-api
* Test naming convention changed to IT and extra test class moved to NonRetryableException test
