+++ pom.xml
@@ -28,12 +28,14 @@
         <private-api-sdk-java.version>4.0.335</private-api-sdk-java.version>
         <kafka-models.version>3.0.20</kafka-models.version>
         <test-containers.version>1.21.3</test-containers.version>
-        <skip.integration.tests>false</skip.integration.tests>
         <mockito-core.version>5.18.0</mockito-core.version>
         <mockito-junit-jupiter.version>5.18.0</mockito-junit-jupiter.version>
         <jib.version>3.4.6</jib.version>
         <maven-surefire-plugin.version>3.5.3</maven-surefire-plugin.version>
         <maven-failsafe-plugin.version>3.5.3</maven-failsafe-plugin.version>
+
+        <skip.unit.tests>false</skip.unit.tests>
+        <skip.integration.tests>false</skip.integration.tests>
 
         <!--sonar configuration-->
         <sonar.token>${CODE_ANALYSIS_TOKEN}</sonar.token>
@@ -157,6 +159,12 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${maven-surefire-plugin.version}</version>
+                <configuration>
+                    <argLine>-Dfile.encoding=UTF-8</argLine>
+                    <argLine>${surefireArgLine}</argLine>
+                    <redirectTestOutputToFile>true</redirectTestOutputToFile>
+                    <skipTests>${skip.unit.tests}</skipTests>
+                </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -164,6 +172,7 @@
                 <version>${maven-failsafe-plugin.version}</version>
                 <executions>
                     <execution>
+                        <phase>integration-test</phase>
                         <goals>
                             <goal>integration-test</goal>
                             <goal>verify</goal>
@@ -171,12 +180,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <!-- See SUREFIRE-1198 -->
-                    <classesDirectory>${project.build.outputDirectory}</classesDirectory>
                     <properties>
-                        <configurationParameters>
-                            cucumber.junit-platform.naming-strategy=long
-                        </configurationParameters>
                         <property>
                             <name>listener</name>
                             <value>org.sonar.java.jacoco.JUnitListener</value>
@@ -184,6 +188,7 @@
                     </properties>
                     <argLine>-Dfile.encoding=UTF-8</argLine>
                     <argLine>${failsafeArgLine}</argLine>
+                    <classesDirectory>${project.build.outputDirectory}</classesDirectory>
                     <skipITs>${skip.integration.tests}</skipITs>
                 </configuration>
             </plugin>
@@ -192,29 +197,25 @@
                 <artifactId>jib-maven-plugin</artifactId>
                 <version>${jib.version}</version>
                 <configuration>
+                    <containerizingMode>packaged</containerizingMode>
                     <from>
-                        <image>416670754337.dkr.ecr.eu-west-2.amazonaws.com/ci-corretto-build-21</image>
+                        <image>416670754337.dkr.ecr.eu-west-2.amazonaws.com/ci-corretto-build-21:latest</image>
                     </from>
                     <to>
-                        <image>416670754337.dkr.ecr.eu-west-2.amazonaws.com/officers-search-consumer</image>
+                        <image>416670754337.dkr.ecr.eu-west-2.amazonaws.com/officers-search-consumer:latest</image>
                     </to>
-                    <container>
-                        <expandClasspathDependencies>true</expandClasspathDependencies>
-                    </container>
                 </configuration>
             </plugin>
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
                 <version>${jacoco-maven-plugin.version}</version>
-
                 <configuration>
                     <append>true</append>
                     <destFile>${sonar.jacoco.reports}/jacoco.exec</destFile>
                     <dataFile>${sonar.jacoco.reports}/jacoco.exec</dataFile>
                     <outputDirectory>${sonar.jacoco.reports}/jacoco</outputDirectory>
                 </configuration>
-
                 <executions>
                     <execution>
                         <id>pre-unit-test</id>
